+++ api/src/main/java/com/tdns/toks/api/domain/user/controller/UserController.java
@@ -4,6 +4,7 @@ import com.tdns.toks.api.domain.user.model.dto.UserApiDTO.UserLoginRequest;
 import com.tdns.toks.api.domain.user.service.UserApiService;
 import com.tdns.toks.core.common.model.dto.ResponseDto;
 import com.tdns.toks.core.common.type.JwtToken;
+import com.tdns.toks.core.domain.user.model.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -12,6 +13,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -40,4 +42,7 @@ public class UserController {
         var response = userApiService.login(userLoginRequest);
         return ResponseDto.ok(response);
     }// 초기세팅 테스트용 메서드
+
+//    @PatchMapping("/name")
+//    public ResponseEntity<User>
 }
