--- core/src/main/java/com/tdns/toks/core/common/oauth/OAuth2SuccessHandler.java
@@ -1,38 +1,0 @@
-package com.tdns.toks.core.common.oauth;
-
-import com.tdns.toks.core.config.security.JwtTokenProvider;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-        log.info("로그인 success");
-
-        // JWT 토큰 발행  및 리다이렉트 지정
-        String url = makeRedirectUrl(jwtTokenProvider.generateToken(authentication.getName()).toString());
-
-        getRedirectStrategy().sendRedirect(request, response, makeRedirectUrl(url));
-    }
-
-    private String makeRedirectUrl(String token) {
-        return UriComponentsBuilder.fromUriString("http://localhost:3000/프로트가원하는주소")
-                .queryParam("token", token)
-                .build().toUriString();
-    }
-}
