+++ core/src/main/java/com/tdns/toks/core/common/security/CustomSessionInformationExpiredStrategy.java
@@ -1,4 +1,4 @@
-package com.tdns.toks.core.config.security;
+package com.tdns.toks.core.common.security;
 
 import com.tdns.toks.core.common.exception.ApplicationErrorException;
 import com.tdns.toks.core.common.exception.ApplicationErrorType;
