+++ core/src/main/java/com/tdns/toks/core/common/security/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.tdns.toks.core.config.security;
+package com.tdns.toks.core.common.security;
 
 import com.google.common.net.HttpHeaders;
 import com.tdns.toks.core.common.type.AuthTokenType;
