+++ core/src/main/java/com/tdns/toks/core/common/security/oauth/OAuth2SuccessHandler.java
@@ -1,0 +1,49 @@
+package com.tdns.toks.core.common.security.oauth;
+
+import com.tdns.toks.core.common.security.JwtTokenProvider;
+import com.tdns.toks.core.common.type.JwtToken;
+import com.tdns.toks.core.domain.user.model.dto.UserDetailDTO;
+import com.tdns.toks.core.domain.user.model.entity.User;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+import org.springframework.web.util.UriComponentsBuilder;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
+import java.util.Map;
+import java.util.Objects;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+
+    private final JwtTokenProvider jwtTokenProvider;
+
+    @Value("${spring.security.oauth2.front-redirect-uri}")
+    private String frontRedirectUri;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        String email = authentication.getName();
+        String url = setRedirectUrl(jwtTokenProvider.generateToken(email));
+        getRedirectStrategy().sendRedirect(request, response, url);
+    }
+
+    private String setRedirectUrl(JwtToken jwtToken) {
+        String accessToken = jwtToken.getAccessToken();
+        String refreshToken = jwtToken.getRefreshToken();
+        return UriComponentsBuilder.fromUriString(frontRedirectUri)
+                .queryParam("accessToken", accessToken)
+                .queryParam("refreshToken", refreshToken)
+                .build().toUriString();
+    }
+}
